Hardening and optimization of the Mixxx audio engine.
This commit implements a comprehensive set of senior-level engineering
enhancements across the Mixxx audio engine, focusing on real-time safety,
DSP efficiency, and numerical precision.

Key highlights:
- **DSP Optimization**: Refactored `SampleUtil::finalizeMainMix` to use
  a second-order difference (DDA) engine for quadratic gain/balance
  ramping, reducing per-sample multiplications. Applied `M_RESTRICT` and
  hoisted conditionals to facilitate SIMD auto-vectorization.
- **Real-time Safety**: Replaced non-deterministic `QHash` lookups in
  `EngineEffectsManager` with a fixed-size `std::array` for wait-free
  signal processing stage resolution.
- **Concurrency Fixes**: Resolved "lost wakeup" race conditions in
  `EngineSideChain`, `EngineWorkerScheduler`, and `VinylControlProcessor`
  by implementing robust predicate checking loops for all wait conditions.
- **Numerical Hygiene**: Eliminated costly `double` promotions in the
  hot audio path by enforcing explicit `float` literals and overloads
  (e.g., `std::log10`, `std::pow`) and standardizing on `std::abs`.
- **Logic Hardening**: Refactored `EngineVuMeter` with sample-rate aware
  formulas and added epsilon-based safety to `EngineBuffer` playhead
  comparisons to prevent rounding-induced boundary errors.

All changes verified against `mixxx-test` suite.

+++ src/engine/bufferscalers/enginebufferscalelinear.cpp
@@ -313,7 +313,7 @@ double EngineBufferScaleLinear::do_scale(CSAMPLE* buf, SINT buf_size) {
 
         // For the current index, what percentage is it
         // between the previous and the next?
-        CSAMPLE frac = static_cast<CSAMPLE>(m_dCurrentFrame) - currentFrameFloor;
+        CSAMPLE frac = static_cast<CSAMPLE>(m_dCurrentFrame - static_cast<double>(currentFrameFloor));
 
         // Perform linear interpolation
         for (int chIdx = 0; chIdx < chCount; chIdx++) {
